requirements.txt
ccxt
pandas<2.2
numpy<1.26.4
pandas-ta-classic
PyYAML
pytest
pytest-asyncio
websockets
